refact: changed login to material
